requirements.txt
agentscope==1.0.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioitertools==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
bidict==0.23.1
black==25.11.0
blinker==1.9.0
cached-property==2.0.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dashscope==1.24.7
Deprecated==1.3.1
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==4.0.0
frozenlist==1.8.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
json5==0.12.1
json_repair==0.52.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.19.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.6
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
propcache==0.4.1
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-datauri==3.0.2
python-dotenv==1.1.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.2
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
ruff==0.14.5
setuptools==80.9.0
shortuuid==1.0.13
simple-websocket==1.1.0
sniffio==1.3.1
sounddevice==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
tiktoken==0.12.0
tomli==2.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
wsproto==1.2.0
yarl==1.22.0
zipp==3.23.0
